+++ src/habitat/IHabitatGameLogic.java
@@ -1,10 +1,8 @@
 package habitat;
 
-import com.sun.java.swing.plaf.windows.WindowsDesktopIconUI;
-
 public interface IHabitatGameLogic {
     void init() throws Exception;
-    void input(int windowHandle);
+    void input(long windowHandle);
     void update(float interval);
-    void render(int windowHandle);
+    void render(long windowHandle);
 }
